Some README fixes. Refactor AWSDynamoDBTable into multiple files to avoid excessively long files. Added travis configuration.
